requirements.txt
aiohttp
log.py
discord
praw
python-dotenv
datetime
psutil
asyncio
discord-pretty-help
discord-paginator
jishaku